pyproject.toml
[project]
name = "b2broker-task"
version = "0.1.0"
description = "Arif Garayev's B2BROKER Take home solution"
authors = [
    {name = "Arif Garayev"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "autoflake (>=2.3.1,<3.0.0)",
    "pytype (>=2024.10.11,<2025.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "flake8-pyproject (>=1.2.3,<2.0.0)",
    "bandit (>=1.8.6,<2.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pylint (>=3.3.7,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django (==5.1)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "djangorestframework (==3.15.0)",
    "model-bakery (>=1.20.5,<2.0.0)",
    "python-environ (>=0.4.54,<0.5.0)",
    "markdown (>=3.8.2,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "tox (>=4.27.0,<5.0.0)",
    "djangorestframework-jsonapi (>=7.1.0,<8.0.0)",
    "psycopg-binary (==3.2.9)",
    "psycopg-pool (>=3.2.6,<4.0.0)",
    "drf-spectacular-jsonapi (>=0.5.2,<0.6.0)"
]

[tool.poetry]
package-mode = false


[tool.isort]
profile = "black"
line_length = 60
py_version = 312
order_by_type = false
skip = [".gitignore", ".dockerignore"]
extend_skip = [".md", ".json"]
skip_glob = ["docs/*"]

[tool.flake8]
max-line-length = 60
select = ["C", "E", "F", "W", "B", "B9"]
ignore = ["E203", "E501", "W503", "C812", "E731", "F811"]
exclude = ["__init__.py"]

[tool.black]
line-length = 60
target-version = ['py312']
include = '\.pyi?$'

[tool.bandit]
skips = ["B311", "B404"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
